packages/a2a-server/pyproject.toml
[project]
name = "mfa-a2a-server"
version = "0.1.0"
description = "A2A server + LangGraph agent runner for Monday-for-Agents"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http]>=0.3.0",
    "langgraph>=0.4.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-anthropic>=0.3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
    "click>=8.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
]

[project.scripts]
mfa = "a2a_server.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/a2a_server"]
